+++ test/rational_precision2double.c
@@ -177,9 +177,9 @@ static void _XTIFFDefaultDirectory(TIFF *tif)
     uint32_t n, nadded;
 
     /* Install the extended Tag field info */
-    n = N(tifFieldInfo);
+    n = (uint32_t)N(tifFieldInfo);
     //_TIFFMergeFields(tif, const TIFFField info[], uint32_t n);
-    nadded = _TIFFMergeFields(tif, tifFieldInfo, n);
+    nadded = (uint32_t)_TIFFMergeFields(tif, tifFieldInfo, n);
     (void)nadded;
 
     /* Since an XTIFF client module may have overridden
@@ -610,7 +610,7 @@ int write_test_tiff(TIFF *tif, const char *filenameRead, int blnAllCustomTags)
         /*==== Automatically check all custom rational tags == WRITING ===*/
         /*-- Get array, where TIFF tag fields are defined --*/
         tFieldArray = _TIFFGetFields();
-        nTags = tFieldArray->count;
+        nTags = (int32_t)tFieldArray->count;
 
         for (i = 0; i < nTags; i++)
         {
@@ -1107,7 +1107,7 @@ int write_test_tiff(TIFF *tif, const char *filenameRead, int blnAllCustomTags)
 
         /*-- Get array, where standard TIFF tag fields are defined --*/
         tFieldArray = _TIFFGetFields();
-        nTags = tFieldArray->count;
+        nTags = (int32_t)tFieldArray->count;
 
         for (i = 0; i < nTags; i++)
         {
@@ -1295,7 +1295,7 @@ int write_test_tiff(TIFF *tif, const char *filenameRead, int blnAllCustomTags)
                                 tSetFieldType == TIFF_SETGET_C32_FLOAT)
                             {
                                 memcpy(&auxFloatArray, pVoidArray,
-                                       (auxInt32 * sizeof(auxFloatArray[0])));
+                                       ((size_t)auxInt32 * sizeof(auxFloatArray[0])));
                                 /* compare read values with written ones */
                                 if (tType == TIFF_RATIONAL ||
                                     tType == TIFF_SRATIONAL)
@@ -1324,7 +1324,7 @@ int write_test_tiff(TIFF *tif, const char *filenameRead, int blnAllCustomTags)
                             else
                             {
                                 memcpy(&auxDoubleArray, pVoidArray,
-                                       (auxInt32 * sizeof(auxDoubleArray[0])));
+                                       ((size_t)auxInt32 * sizeof(auxDoubleArray[0])));
                                 /* compare read values with written ones */
                                 if (tType == TIFF_RATIONAL ||
                                     tType == TIFF_SRATIONAL)
